+++ crates/vela_engine/src/standard/methods/map.rs
@@ -137,6 +137,6 @@ const MAP_METHODS: &[MethodSpec] = &[
     ),
 ];
 
-const fn map_id(offset: u32) -> HostMethodId {
+const fn map_id(offset: u64) -> HostMethodId {
     HostMethodId::new(0xff00_0900 + offset)
 }
